refactor(api): consolidate hand-written handler helpers in httphelpers.go
writeJSON / writeProblem / pathUUID (previously in
cloud_account_handlers.go) and requireScope / parseLimit (previously in
application_block_handlers.go) move to a new internal/api/httphelpers.go.

The 16 handlers that re-implemented the scope-guard inline (extracts,
virtual machines, cloud accounts, security-group sweep) now call
requireScope; sites that use the caller afterwards fetch it after the
guard. Strict-server handlers (network policies) keep their typed 403
responses. Problem-JSON bodies are byte-for-byte unchanged
(scope constants match the previous literal messages).
